+++ src/res/listeners.js
@@ -1,3 +1,3 @@
-export const GLOBAL_LISTENER_NAME = '__global__'
+export const GLOBAL_NAMESPACE = '__global__'
 export const ON_RESOLVED = 'resolved'