+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Unreleased]
+### Added
+  - Decorator functionality added (#8)
+  - Parameter section, in configuration object, now accepts objects to better customization of parameters at configuration level.
 
 ## [1.1.1] - 2016-10-21
 ### Added